// src/app/models/response.model.ts
export class ObsResponse {
	obsResponse: string;
	type: string;
	message: string;
	attachmens: any[];
	reportCycle: string;
	isEditable: boolean;
	completionYear: string;
	stepUniqueName: string;
	approveEnabled: any;
	obsType: string;
	govtEntity: string;
	status?:string;
	prevCompletionYear?:string;
	prevGovtEntity?:string;
	sabRequest?:string;
	sabCommentary?:string;
}